+++ src/main/java/ua/com/zaibalo/db/api/MessagesDAO.java
@@ -7,7 +7,7 @@ import ua.com.zaibalo.model.Message;
 public interface MessagesDAO {
 	int insert(Message message);
 	List<Message> getAllUserDiscussionMessages(int discussionId, int userId);
-	int getUnreadMessagesCount(int recipientId);
+	long getUnreadMessagesCount(int recipientId);
 	void setDialogMessagesRead(int discussionId, int recipientId);
 	Message getMessageById(int messageId);
 }
